Add arduino libraries, add function declarations
Addded the used arduino libraries to the makefile so arduino-mk is able
to read the .h files needed. Also added the function declarations to
hind-brain.h because make does not do the preprocessing that the arduino
ide normally handles. This is now in a working state, to build and
upload, `make` and `make upload` can be used