Update vel & steering limits; refactor limit checking
1. Update steer & vel tuning constants to recent measured vals
2. Add auto-calculation of MAX & MIN signals for steer & vel - the
result of this is to make safety thresholding in ackToCmd() functions
more resilient to flipping the potentiometer. Safety thresholding code
now compares MAX and MIN rather than HIGH and LOW.
3. Separate range conversion for velocity to match steering - the result
of this change is to allow a centerpoint to be used that is not midway
between the FWD and REV values - necessary now that the vehicle can move
in reverse.